+++ scripts/test/Selenium/Agent/AgentTicketSearch.t
@@ -249,6 +249,38 @@ $Selenium->RunTest(
             "Ticket $TitleRandom found on page with correct StaticDB search",
         );
 
+        # navigate to AgentTicketSearch screen again
+        $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AgentTicketSearch");
+
+        # wait until form and overlay has loaded, if neccessary
+        $Selenium->WaitFor( JavaScript => "return typeof(\$) === 'function' && \$('#SearchProfile').length" );
+
+        # add search filter by priority and run it
+        $Selenium->execute_script(
+            "\$('#Attribute').val('PriorityIDs').trigger('redraw.InputField').trigger('change');",
+        );
+        $Selenium->find_element( '.AddButton',          'css' )->click();
+        $Selenium->find_element( '#PriorityIDs_Search', 'css' )->click();
+
+        # wait until drop down list is shown
+        $Selenium->WaitFor(
+            JavaScript => "return typeof(\$) === 'function' && \$('.InputField_ListContainer').length"
+        );
+
+        # click on remove button next to priority field
+        $Selenium->find_element( '#PriorityIDs + .RemoveButton', 'css' )->click();
+
+        # wait until drop down list is hidden
+        $Selenium->WaitFor(
+            JavaScript => "return typeof(\$) === 'function' && \$('.InputField_ListContainer').length == 0"
+        );
+
+        # verify dropdown list has been hidden (bug#12243)
+        $Self->True(
+            index( $Selenium->get_page_source(), 'InputField_ListContainer' ) == -1,
+            "InputField list not found on page",
+        );
+
         # clean up test data from the DB
         my $Success = $TicketObject->TicketDelete(
             TicketID => $TicketID,
@@ -262,7 +294,7 @@ $Selenium->RunTest(
         # make sure the cache is correct
         $Kernel::OM->Get('Kernel::System::Cache')->CleanUp( Type => 'Ticket' );
 
-    }
+    },
 );
 
 1;